Fixing transformer crash on Null Arrays
Redis return Null Arrays in specific situations (e.g. BLPOP
timeouts or EXEC if aborting).

Before this commit, _MultiBulkConsumer crashed with a
RangeError, while trying to create a List with -1 lenght.

Redis doc states: "A client library API should return a null
object and not an empty Array when Redis replies with a
Null Array".
This simple fix does not comply with this suggestion. A larger
rework would be required to return a Null object.